:recycle: move logic from camera loader to reference loader
